requirements.txt
cnocr==2.2.4.2
opencv_python==4.8.1.78
Pillow==10.0.1
Pillow==10.1.0
PyAutoGUI==0.9.54
pyperclip==1.8.2
Requests==2.31.0
translate==3.6.1
